Extract `parseConnectionToken` to a separate file
